requirements.txt
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
python-dotenv==1.0.1
djangorestframework-simplejwt==5.5.0